+++ app/assets/javascripts/bootsy/init.js
@@ -1,13 +1,15 @@
 window.Bootsy = window.Bootsy || {};
 
 Bootsy.init = function() {
-  Bootsy.areas = [];
+  Bootsy.areas = {};
 
-  $('textarea.bootsy_text_area').each(function() {
+  $('textarea.bootsy_text_area').each(function(index) {
     var area = new Bootsy.Area($(this));
+    var areaIdx = $(this).attr('id') || index;
+
     area.init();
 
-    Bootsy.areas.push(area);
+    Bootsy.areas[areaIdx] = area;
   });
 };